Fix Direct2DHwndContext incorrectly reporting DPI scaling

+++ modules/juce_graphics/native/juce_Direct2DHwndContext_windows.cpp
@@ -345,7 +345,7 @@ public:
         return hr;
     }
 
-    void setScaleFactor (double scale_) override
+    void setScaleFactor (float scale_) override
     {
         Pimpl::setScaleFactor (scale_);
